feat: change session logic display
