+++ packages/core/src/auth/i18n/fr.json
@@ -14,7 +14,7 @@
   "Auth_Change_Session": "Changer de session",
   "Auth_Saved_Sessions": "Sessions enregistrées",
   "Auth_Session_Name_Aleary_Exist": "Ce nom de session existe déjà",
-  "Auth_Create_Session": "créer une session",
+  "Auth_Create_Session": "Créer une session",
   "Auth_InfoSession": "Pour vous connecter à l'application mobile, vous devez créer une session en indiquant l'url de votre instance ainsi que votre nom d'utilisateur et votre mot de passe. Une fois que vous aurez créé votre première session, vous pourrez vous connecter plus facilement en saisissant uniquement le mot de passe. Vous pouvez également gérer un ensemble de sessions.",
   "User_User": "Utilisateur",
   "User_VirtualKeyboardConfig": "Cacher le clavier virtuel",
