+++ packages/core/src/components/templates/index.js
@@ -20,3 +20,4 @@ export {default as AttachedFilesView} from './AttachedFilesView/AttachedFilesVie
 export {default as MailMessageView} from './MailMessageView/MailMessageView';
 export {default as PopupCreateSession} from './PopupCreateSession/PopupCreateSession';
 export {default as PopupSession} from './PopupSession/PopupSession';
+export {default as PopupSessionList} from './PopupSessionList/PopupSessionList';
